borgc: widen SPIR-B const_vals wire format from u16 to u32
Companion to the Borg-side widening (borg_spirb.h/c, docs/spirb.md, same
day, feat/fp32-datapath branch): const_vals is now emitted as LE u32
instead of LE u16, so the wire format can eventually carry a real FP32
constant.

Wire-format-only for now -- borgc has no FP32 codegen path (no notion
of a target float format anywhere in this compiler), so every value
emitted today is still f32_to_fp16(bits) as u32: an FP16 bit pattern
zero-extended into the wider slot, not a real FP32 constant. Actually
compiling shaders to FP32 is a separate, larger undertaking (NIR
lowering, register allocation, instruction selection all becoming
float-format-aware) tracked separately, not attempted here.

Verified: Borg-side software/borg native test suite (test_spirb.c,
15/15) against the matching parser change. This compiler crate itself
was NOT rebuilt via the full mesa/meson build (no configured build
directory in this fresh submodule checkout, and setting one up was out
of proportion for a 3-line, purely mechanical type-widening change) --
verified instead by manual review: emit_blob's signature, its one call
site, and its one const_uniforms.push() call site are the only three
touch points, all type-consistent. Flagging this gap rather than
silently claiming full build verification.

+++ src/borg/compiler/lib.rs
@@ -186,7 +186,10 @@ pub unsafe extern "C" fn borgc_compile_nir(
     // the full pipeline, not the fragment in isolation.)  Uniforms are too tight:
     // rast(u0-11) + frag varyings(u12-30) already fill 31 of 32. Records (reg, fp16).
     let mut const_reg_next: u8 = 17;
-    let mut const_uniforms: Vec<(u8, u16)> = Vec::new();
+    // u32 (widened from u16): wire-format-only, see emit_blob's doc comment
+    // in encode.rs -- every value pushed today is still an fp16 bit pattern
+    // zero-extended via `as u32`, not a real FP32 constant.
+    let mut const_uniforms: Vec<(u8, u32)> = Vec::new();
     // gl_varying_slot: VAR0=texcoord, VAR1=frag_pos (Mesa enum: VAR0 = 32).
     const VARYING_SLOT_VAR0: u32 = 32;
     if !entry.is_null() {
@@ -404,7 +407,7 @@ pub unsafe extern "C" fn borgc_compile_nir(
                                 let reg = const_reg_next;
                                 const_reg_next += 1;
                                 let bits = unsafe { lc.values()[c].u32_ };
-                                const_uniforms.push((reg, f32_to_fp16(bits)));
+                                const_uniforms.push((reg, f32_to_fp16(bits) as u32));
                                 let v = next_vreg;
                                 next_vreg += 1;
                                 ubo.insert(v, Ubo::Fixed(reg));
